backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
transformers==4.36.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.1+cpu
python-multipart==0.0.9
pydantic==2.6.1
accelerate==0.25.0 